refactor:shift main page folder
