+++ src/pages/main-page/main-page.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { PageProps } from '../../../interfaces/pages.intarface';
+import { PageProps } from '../../interfaces/pages.intarface';
 
 const MainPage = ({ showName }: PageProps) => {
   useEffect(() => {
